pyproject.toml
[project]
name = "horizon_orchestrator"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Savvin", email = "nikita.vsav@gmail.com"},
    {name = "Dmitry Vasenin", email = "vasenindmitrij01@gmail.com"},
    {name = "Anastasia Shahova", email = "anastasia.shakhova1@gmail.com"},
    {name = "Nikita Ryzhenkov", email = "nikitar2025@mail.ru"},
]
dependencies = [
    "pytest>=8.1.1",
    "tox==4.14.2",
    "pre-commit>=3.6.2",
    "pytest-cov>=4.1.0",
    "coverage-badge>=1.1.0",
    "pdm-bump>=0.7.3",
    "pydantic>=2.6.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.28.0",
    "coverage==7.4.4",
    "environs>=14.3.0",
    "ruff>=0.12.7",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.7.1",
    "pandas>=2.3.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "fastapi-cprofile>=0.0.2",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pytest.ini_options]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]
timeout = 300
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit_transform",
]
pythonpath = [
    ".",
]

[tool.ruff]

exclude = [
    "tests/",
    "*.ipynb",
    "__pycache__",
]

line-length = 120
select = [
  "F",  # pyflakes
  "I",  # isort 
]

ignore = ["E501", "F811"]  # Игнорировать длину строки и неиспользуемые функции